+++ code/src/sixsq/nuvla/server/resources/credential_template.clj
@@ -123,7 +123,7 @@ curl https://nuvla.io/api/credential-template
 ;; must override the default implementation so that the
 ;; data can be pulled from the atom rather than the database
 (defmethod crud/retrieve-by-id resource-type
-  [id]
+  [id & _]
   (try
     (get @templates id)
     (catch Exception e