+++ code/src/sixsq/nuvla/server/resources/group_template.clj
@@ -101,7 +101,7 @@ creating a group and does not provide any useful defaults.
 ;; must override the default implementation so that the
 ;; data can be pulled from the atom rather than the database
 (defmethod crud/retrieve-by-id resource-type
-  [id]
+  [id & _]
   (try
     (get @templates id)
     (catch Exception e