+++ code/src/sixsq/nuvla/server/resources/resource_metadata.clj
@@ -100,7 +100,7 @@ attribute.
 ;; must override the default implementation so that the
 ;; data can be pulled from the atom rather than the database
 (defmethod crud/retrieve-by-id resource-type
-  [id]
+  [id & _]
   (try
     (get @templates id)
     (catch Exception e