+++ CHANGELOG.md
@@ -8,11 +8,12 @@
 
 ### Changed
 
+  - NuvlaBox Status - add optional attribute for temperatures
   - Retrieve-by-id check acl
   - Event - Enhancement
   - Deployment - Bulk-force-delete
-  - Deployment - force-delete action 
-  - Deployment - Bulk-stop 
+  - Deployment - force-delete action
+  - Deployment - Bulk-stop
   - Module - Fix boolean type spec
   - Spec-tools - Fix broken changes
   - Dependencies - Update sixsq.nuvla/parent 6.7.5